Fix bug with import and cleanup
