+++ node/db/tests/mem_test.rs
@@ -6,14 +6,6 @@ mod subtests;
 use db::MemoryDB;
 
 #[test]
-fn mem_db_open() {
-    let mut db = MemoryDB::default();
-    subtests::open(&mut db);
-    // Calling open on opened db should not error
-    subtests::open(&mut db);
-}
-
-#[test]
 fn mem_db_write() {
     let db = MemoryDB::default();
     subtests::write(&db);
