+++ node/db/tests/rocks_test.rs
@@ -10,19 +10,10 @@ use db::RocksDb;
 use db_utils::DBPath;
 
 #[test]
-fn rocks_db_open() {
-    let path = DBPath::new("start_rocks_test");
-    let mut db = RocksDb::new(path.as_ref());
-    subtests::open(&mut db);
-    // Calling open on opened db should not error
-    subtests::open(&mut db);
-}
-
-#[test]
 fn rocks_db_write() {
     let path = DBPath::new("write_rocks_test");
     let mut db = RocksDb::new(path.as_ref());
-    subtests::open(&mut db);
+    db.open().unwrap();
     subtests::write(&db);
 }
 
@@ -30,7 +21,7 @@ fn rocks_db_write() {
 fn rocks_db_read() {
     let path = DBPath::new("read_rocks_test");
     let mut db = RocksDb::new(path.as_ref());
-    subtests::open(&mut db);
+    db.open().unwrap();
     subtests::read(&db);
 }
 
@@ -38,7 +29,7 @@ fn rocks_db_read() {
 fn rocks_db_exists() {
     let path = DBPath::new("exists_rocks_test");
     let mut db = RocksDb::new(path.as_ref());
-    subtests::open(&mut db);
+    db.open().unwrap();
     subtests::exists(&db);
 }
 
@@ -46,7 +37,7 @@ fn rocks_db_exists() {
 fn rocks_db_does_not_exist() {
     let path = DBPath::new("does_not_exists_rocks_test");
     let mut db = RocksDb::new(path.as_ref());
-    subtests::open(&mut db);
+    db.open().unwrap();
     subtests::does_not_exist(&db);
 }
 
@@ -54,7 +45,7 @@ fn rocks_db_does_not_exist() {
 fn rocks_db_delete() {
     let path = DBPath::new("delete_rocks_test");
     let mut db = RocksDb::new(path.as_ref());
-    subtests::open(&mut db);
+    db.open().unwrap();
     subtests::delete(&db);
 }
 
@@ -62,7 +53,7 @@ fn rocks_db_delete() {
 fn rocks_db_bulk_write() {
     let path = DBPath::new("bulk_write_rocks_test");
     let mut db = RocksDb::new(path.as_ref());
-    subtests::open(&mut db);
+    db.open().unwrap();
     subtests::bulk_write(&db);
 }
 
@@ -70,7 +61,7 @@ fn rocks_db_bulk_write() {
 fn rocks_db_bulk_read() {
     let path = DBPath::new("bulk_read_rocks_test");
     let mut db = RocksDb::new(path.as_ref());
-    subtests::open(&mut db);
+    db.open().unwrap();
     subtests::bulk_read(&db);
 }
 
@@ -78,6 +69,6 @@ fn rocks_db_bulk_read() {
 fn rocks_db_bulk_delete() {
     let path = DBPath::new("bulk_delete_rocks_test");
     let mut db = RocksDb::new(path.as_ref());
-    subtests::open(&mut db);
+    db.open().unwrap();
     subtests::bulk_delete(&db);
 }
